refactor(support): repackage views
