+++ src/main/kotlin/apply/ui/admin/selections/SelectionView.kt
@@ -16,11 +16,11 @@ import com.vaadin.flow.router.BeforeEvent
 import com.vaadin.flow.router.HasUrlParameter
 import com.vaadin.flow.router.Route
 import com.vaadin.flow.router.WildcardParameter
-import support.addSortableColumn
-import support.addSortableDateColumn
-import support.createPrimarySmallButton
-import support.createSearchBar
-import support.createSuccessButton
+import support.views.addSortableColumn
+import support.views.addSortableDateColumn
+import support.views.createPrimarySmallButton
+import support.views.createSearchBar
+import support.views.createSuccessButton
 
 @Route(value = "admin/selections", layout = BaseLayout::class)
 class SelectionView(
