+++ src/main/kotlin/support/views/Buttons.kt
@@ -1,26 +1,13 @@
-package support
+package support.views
 
 import com.vaadin.flow.component.ClickEvent
-import com.vaadin.flow.component.ComponentEventListener
-import com.vaadin.flow.component.Key
-import com.vaadin.flow.component.KeyDownEvent
 import com.vaadin.flow.component.Text
 import com.vaadin.flow.component.UI
 import com.vaadin.flow.component.button.Button
 import com.vaadin.flow.component.button.ButtonVariant
 import com.vaadin.flow.component.dialog.Dialog
-import com.vaadin.flow.component.grid.Grid
-import com.vaadin.flow.component.html.Div
-import com.vaadin.flow.component.icon.Icon
-import com.vaadin.flow.component.icon.VaadinIcon
 import com.vaadin.flow.component.orderedlayout.FlexComponent
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout
-import com.vaadin.flow.component.select.Select
-import com.vaadin.flow.component.textfield.TextField
-import com.vaadin.flow.data.renderer.LocalDateRenderer
-import com.vaadin.flow.data.renderer.LocalDateTimeRenderer
-import java.time.LocalDate
-import java.time.LocalDateTime
 
 typealias ClickListener = (ClickEvent<Button>) -> Unit
 
@@ -84,7 +71,7 @@ fun createDeleteButtonWithDialog(message: String, clickListener: ClickListener):
     return createErrorSmallButton("삭제") { createConfirmDialog(message, clickListener).open() }
 }
 
-fun createConfirmDialog(
+private fun createConfirmDialog(
     text: String,
     confirmListener: ClickListener,
     cancelListener: ClickListener = {}
@@ -113,52 +100,4 @@ private fun Dialog.createConfirmButton(clickListener: ClickListener): Button {
         clickListener(it)
         UI.getCurrent().page.reload()
     }
-}
-
-fun createIntSelect(min: Int = 0, max: Int): Select<Int> {
-    return Select(*(min..max).toList().toTypedArray())
-}
-
-fun <T : Any> Grid<T>.addSortableColumn(labelText: String, valueProvider: (T) -> Any): Grid.Column<T> {
-    return addColumn(valueProvider).apply {
-        addSortableHeader(labelText)
-    }
-}
-
-fun <T : Any> Grid<T>.addSortableDateTimeColumn(
-    labelText: String,
-    valueProvider: (T) -> LocalDateTime
-): Grid.Column<T> {
-    return addColumn(LocalDateTimeRenderer(valueProvider, "yyyy-MM-dd HH:mm:ss")).apply {
-        addSortableHeader(labelText)
-        setComparator(compareBy(valueProvider))
-    }
-}
-
-fun <T : Any> Grid<T>.addSortableDateColumn(
-    labelText: String,
-    valueProvider: (T) -> LocalDate
-): Grid.Column<T> {
-    return addColumn(LocalDateRenderer(valueProvider, "yyyy-MM-dd")).apply {
-        addSortableHeader(labelText)
-        setComparator(compareBy(valueProvider))
-    }
-}
-
-private fun <T : Any> Grid.Column<T>.addSortableHeader(labelText: String) {
-    setHeader(labelText)
-    isSortable = true
-    isAutoWidth = true
-}
-
-fun createSearchBar(eventListener: (name: String) -> Unit): Div {
-    val textField = TextField()
-    textField.addKeyDownListener(
-        Key.ENTER,
-        ComponentEventListener<KeyDownEvent?> { eventListener(textField.value) }
-    )
-    return Div(
-        textField,
-        Button(Icon(VaadinIcon.SEARCH)) { eventListener(textField.value) }
-    )
 }
